requirements.txt
libigl
polyscope
numpy
scipy
matplotlib
scikit-learn
opencv-python
cvxopt